requirements.txt
streamlit
google-generativeai
edge-tts
SpeechRecognition
streamlit-audiorecorder
python-dotenv
fuzzywuzzy
python-Levenshtein
numpy
soundfile
pydub
requests
